% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_resp_support.R
\name{classify_resp_support}
\alias{classify_resp_support}
\title{classify_resp_support}
\usage{
classify_resp_support(
  df_resp_wide,
  min_ep_duration = 1,
  max_inter_ep_duration = 4
)
}
\arguments{
\item{df_resp_wide}{A data frame. It must be in wide format and contain
encounter IDs (named enc_id) and different respiratory parameters. The list
of parameters is fixed. This data frame is usually returned by \link{clean_resp_support}.}

\item{min_ep_duration}{A number representing the minimum time (in hours) that
an episode  of respiratory support needs to be in effect for it to be counted.
Respiratory support episodes with duration under this level will be discarded.
Default is 1 hour.}

\item{max_inter_ep_duration}{A number representing the maximum number of hours
in between episodes of respiratory support that will be tolerated. Default is
4 hours. Respiratory therapists usually document every 4 hours at a minimum.
Each episode has a duration (the minimum allowed is defined by min_ep_duration).
If the time between two neighboring episodes is fewer hours than
max_inter_ep_duration, this amount of time will be added to the duration of
an episode. Otherwise, the duration of an episode will only be the time
between when the episode starts and stops. For example, if a patient is on CPAP
for 12 hours, and the next recorded value is 2 hours later and represents
switching to BiPAP, the "duration" of CPAP will be recorded as 14 hours.
However, if the same patient didn't have a new recorded value for 18 hours
representing BiPAP, the "duration" of CPAP will be recorded as just 12 hours.
The time between the end of CPAP and the beginning of BiPAP will be undefined.}
}
\value{
A data frame in long format encounter IDs, start and stop times of
respiratory support, and the duration of time
}
\description{
Determine different levels of respiratory support based upon a wide-format
dataset of encounter IDs and variables dictating different respiratory
flowsheet parameters
}
