% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_vent.R
\name{clean_vent}
\alias{clean_vent}
\title{clean_vent}
\usage{
clean_vent(df_vent)
}
\arguments{
\item{df_vent}{A long-form data frame of vent/respiratory settings and measurements}
}
\value{
A data frame in wide format with concurrently-recorded respiratory data
}
\description{
Take ventilator/respiratory file loaded by \link{load_vent}. Clean it a little,
keep standard valuable rows, and put it into wide format so that each timestamp has
all concurrently-recorded values.
}
