% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_blank_series.R
\name{create_blank_series}
\alias{create_blank_series}
\title{create_blank_series}
\usage{
create_blank_series(id, starttime, stoptime, increment = "hours")
}
\arguments{
\item{id}{A unique character identifier for each patient.}

\item{starttime}{A datetime identifying the beginning of the time series.}

\item{stoptime}{A datetime identifying the end of the time series}

\item{increment}{A character specifying how often to increment the series.
Default is 'hours'. Can use anything that \code{base::seq()} would accept, such as 'mins',
'days', etc.}
}
\value{
A long-format data frame consisting of IDs and timestamps. The timestamp
column can be renamed outside of this function to be used in joins.
}
\description{
Send in ID keys, each with a start and stop datetime. This will create a long-format
data frame of blank values. This can be used in a number of different ways with
sparse EHR data. An example is to create a by-hour dataset of values for the entirety
of a patient's hospitalization. You can join a sparse, timestamped array with
a blank array created here. You can then choose different methods, such as last-one
carried forward (LOCF), to account for missing data on an hourly basis.
}
