% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_capd_intervals.R
\name{get_capd_intervals}
\alias{get_capd_intervals}
\title{get_capd_intervals}
\usage{
get_capd_intervals(
  id,
  capd,
  capd_time,
  coma_times = NULL,
  max_inter_ep_duration = 12
)
}
\arguments{
\item{id}{A character vector of IDs for each unique patient or encounter. Generally taken from
the variable PAT_ENC_CSN_ID, but can also create custom IDs. An example would be to use
PAT_ENC_CSN_ID but also append an identifier (01, 02, 03, etc.) for each PICU hospitalization
or episode of invasive mechanical ventilation.}

\item{capd}{A vector of integers ranging from 0 to 32, representing the CAPD for each patient.}

\item{capd_time}{A vector of datetime or POSIXct entries corresponding to each CAPD value}

\item{coma_times}{A data frame containing 3 columns: id, POSIXct objects corresponding to start/stop times
when RASS is -4 or -5. CAPD is undefined at this time. If \code{NULL} then this will be skipped.
This data frame can be generated by the function \link{get_rass_intervals} with a little processing.}

\item{max_inter_ep_duration}{The maximum number of hours that a particular CAPD will be
extended in the event of a null recording. The default is 12 hours. CAPD may be sparsely recorded.
For any recorded CAPD, the current CAPD will be carried forward for \code{max_inter_ep_duration} hours.
After this amount of time, the interval will be ended. Set this variable to 0 and no intervals will
be created. Set it to NA and the interval will be extended forward indefinitely until a new CAPD is recorded.}
}
\value{
A data frame with:
\itemize {
\item \code{id}: The ID of the patient.
\item \code{capd_episode}: The number of each CAPD interval (1, 2, 3, ..., k) for k intervals per ID
\item \code{capd}: The CAPD
\item \code{capd_time_start}: The start datetime of each interval
\item \code{capd_time_stop}: The end datetime of each interval
\item \code{capd_duration}: A \code{\link[lubridate:duration]{lubridate::duration()}} object representing the duration of this interval
\item \code{delirious}: A logical representing whether the patient is or isn't delirious at this time
}
}
\description{
Takes data fram from \link{load_capd} and creates intervals for each patient encounter.
Each interval contains the Cornell Assessment of Pediatric Delirium (CAPD) score,
as well as the duration of time spent at that interval.
Some processing is incorporated to take comatose times into account, during which
a CAPD cannot be scored.
}
\details{
An interval will last at most max_ep_duration if nothing else is charted.
}
