% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_imv_startstop.R
\name{get_imv_startstop}
\alias{get_imv_startstop}
\title{get_imv_startstop}
\usage{
get_imv_startstop(df_vent_wide, min_imv_time = 2)
}
\arguments{
\item{df_vent_wide}{A wide-format data frame of vent/respiratory settings and measurements. This should
be obtained using \link{clean_vent}}

\item{min_imv_time}{The minimum duration of time (in hours) an episode of IMV must last for it to be counted.
This is designed to make sure that erroneous entries that would flag a patient as extubated will be removed.
Neighboring valid episodes will be joined together.}
}
\value{
A data frame in long format with start and stop times of IMV for each patient/encounter
}
\description{
Find times where invasive mechanical ventilation (IMV) starts and stops, to determine
the exact interval and duration of IMV.
}
