% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lab_type.R
\name{get_lab_by_type}
\alias{get_lab_by_type}
\title{get_lab_by_type}
\usage{
get_lab_by_type(df_labs, labnames = NULL, labvarnames = NULL)
}
\arguments{
\item{df_labs}{A data frame of lab values}

\item{labnames}{Character vector of the lab name(s) that you want to filter on.}

\item{labvarnames}{Character vector that we will use to name columns.

\code{labnames} and \code{labvarnames} must have the same length.}
}
\value{
A data frame of filtered lab values, in wide format at each timepoint.
}
\description{
Utility function. Returns all labs of a given type or types.
}
