% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nicu_encounters.R
\name{get_nicu_encounters}
\alias{get_nicu_encounters}
\title{get_nicu_encounters}
\usage{
get_nicu_encounters(
  adt_filepath,
  adt_coltypes = list(col_character(), col_character(), col_number(), col_character(),
    col_datetime(), col_skip(), col_character(), col_skip(), col_skip(), col_skip()),
  max_load = Inf
)
}
\arguments{
\item{adt_filepath}{Path to the ADT file.}

\item{adt_coltypes}{A list of cols() specifications.
Cols specifications are things like col_integer(), col_character(), and can be found within the 'readr' package documentation.}

\item{max_load}{A number, the maximum number of rows to load. The default is infinity.}
}
\description{
Load an ADT dataset and determine whether a patient was ever in the NICU
}
