% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_or_times.R
\name{get_or_times}
\alias{get_or_times}
\title{get_or_times}
\usage{
get_or_times(
  adt_filepath,
  adt_coltypes = list(col_character(), col_character(), col_number(), col_character(),
    col_datetime(), col_skip(), col_character(), col_skip(), col_skip(), col_skip()),
  max_load = Inf
)
}
\arguments{
\item{adt_filepath}{Path to the ADT file.}

\item{adt_coltypes}{A list of cols() specifications.
Cols specifications are things like col_integer(), col_character(), and can be found within the 'readr' package documentation.}

\item{max_load}{A number, the maximum number of rows to load. The default is infinity.}
}
\value{
#' @return A data frame with:
\itemize{
\item \code{mrn}: Medical record number
\item \code{enc_id}: Encounter ID, renamed from PAT_ENC_CSN_ID
\item \code{or_start_date}: Datetime for the start of each OR course
\item \code{or_stop_date}: Datetime for the end of each OR course
}
}
\description{
Read OR times from the ADT data. Returns times when patients entered and left the OR, IR, or other operative locations.
}
