% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_picu_intervals.R
\name{get_picu_intervals}
\alias{get_picu_intervals}
\title{get_picu_intervals}
\usage{
get_picu_intervals(
  adt_filepath,
  adt_coltypes = list(col_character(), col_character(), col_number(), col_character(),
    col_datetime(), col_skip(), col_character(), col_skip()),
  max_load = Inf
)
}
\arguments{
\item{adt_filepath}{Path to the ADT file.}

\item{adt_coltypes}{A list of cols() specifications.
Cols specifications are things like col_integer(), col_character(), and can be found within the 'readr' package documentation.}

\item{max_load}{A number, the maximum number of rows to load. The default is infinity.}
}
\value{
A data frame with:
\itemize{
  \item \code{mrn}: Medical record number
  \item \code{enc_id}: Encounter ID, renamed from PAT_ENC_CSN_ID
  \item \code{icu_start_date}: Datetime for the start of each ICU stay
  \item \code{icu_stop_date  me}: Datetime for the stop of each ICU stay
}
}
\description{
Load an ADT dataset and determine when a patient entered/left the PICU
}
