% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_config.R
\name{load_config}
\alias{load_config}
\title{load_config}
\usage{
load_config(yml_path = "config/config.yml", useglobal = FALSE)
}
\arguments{
\item{yml_path}{A character vector path to the configuration file. Default is config/config.yml}

\item{useglobal}{A boolean, default to FALSE. Set to TRUE if you want to automatically load configuration variables into the global environment. Otherwise these will be returned as a list.}
}
\value{
Returns nothing if useglobal=TRUE. Otherwise returns a list containing configuration variables.
These can run using \code{list2env(config, globalenv())}
}
\description{
Load basic configuration variables, such as file paths, to be used with
various functions in chonyepicdata
}
