% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_gcs.R
\name{load_gcs}
\alias{load_gcs}
\title{load_gcs}
\usage{
load_gcs(
  gcs_filepath,
  gcs_coltypes = list(col_character(), col_character(), col_skip(), col_character(),
    col_skip(), col_integer(), col_skip(), col_skip(), col_datetime()),
  max_load = Inf
)
}
\arguments{
\item{gcs_filepath}{Path to the file.}

\item{gcs_coltypes}{A list of cols() specifications.
Cols specifications are things like col_integer(), col_character(), and can be found
within the \code{\link[readr]{cols}} documentation from the \code{readr} package.
If this isn't working well you can send in col_guess().}

\item{max_load}{Maximum number of rows}
}
\value{
A timestamped data frame with GCS subtypes and overall score. If data are NA, they are counted as zero.
They are also kept so that we can check. GCS verbal is allowed to be NA when the patient is intubated, but not
otherwise. There is no way to check this in this file, so we will need to cross-reference with output from
\link{classify_resp_support}.
}
\description{
Loads Glasgow Coma Scores for all patients from a raw, pipe-delimited file.
}
