% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_generic_flowsheet_rows.R
\name{load_generic_flowsheet_rows}
\alias{load_generic_flowsheet_rows}
\title{load_generic_flowsheet_rows}
\usage{
load_generic_flowsheet_rows(
  flowsheet_filepath,
  key_name = "pat_enc_csn_id",
  time_col = NULL,
  var_col = NULL,
  measure_col = NULL,
  varnames = NULL,
  rename_vars = NULL,
  max_load = Inf,
  var_transform = NULL
)
}
\arguments{
\item{rename_vars}{}
}
\value{
A data frame with the identity key and timestamped data points from the file.
}
\description{
Send in a flowsheet file of an undefined type, and pull out the relevant rows
in a wide, timestamped format.
}
