% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_icd_dx.R
\name{load_icd_dx}
\alias{load_icd_dx}
\title{load_icd_dx}
\usage{
load_icd_dx(
  df_encounters,
  icd_dx_filepath,
  coltypes_icd = list(col_character(), col_character(), col_character(), col_character(),
    col_datetime(), col_character(), col_factor()),
  max_load = Inf
)
}
\arguments{
\item{df_encounters}{A data frame consisting of MRNs, admission dates "hospital_admission"date, discharge dates "hospital_discharge date"}

\item{icd_dx_filepath}{Path to the ICD diagnosis code file.}

\item{coltypes_icd}{A list of cols() specifications.
Cols specifications are things like col_integer(), col_character(), and can be found within the 'readr' package documentation.}

\item{max_load}{A number, the maximum number of rows to load. The default is infinity.}
}
\value{
A data frame with MRN, encounter ID, ICD10 code and description, dates of code entry, whether the diagnosis was primary (Y/N), and whether the diagnosis was made before, during, or the hospital encounter.
}
\description{
After loading encounter data with function load_encounters(), use this function
to find all ICD diagnosis codes associated with each of these patients. ICD codes will be
loaded from many different sources, and across time. Because of this, also create
a column specifying whether the ICD diagnosis originated before, during, or after the
hospital encounter.
}
