% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_meds.R
\name{load_meds}
\alias{load_meds}
\title{load_meds}
\usage{
load_meds(
  med_filepath,
  med_coltypes = list(col_number(), col_skip(), col_character(), col_character(),
    col_datetime(), col_character(), col_character(), col_double(), col_character(),
    col_character(), col_double(), col_character(), col_character(), col_datetime(),
    col_skip(), col_character()),
  max_load = Inf
)
}
\arguments{
\item{med_filepath}{Path to the medication data}

\item{med_coltypes}{A list of cols() specifications.
Cols specifications are things like col_integer(), col_character(), and can be found
within the \code{\link[readr]{cols}} documentation from the \code{readr} package.
If this isn't working well you can send in col_guess().}

\item{max_load}{The maximum number of rows to load. The default is \code{Inf}}
}
\value{
A data frame with raw medication data. Medication name is saved in character format
in the column 'med_name'. These should be cleaned and processed.
}
\description{
Load medication data from the Epic EHR, and perform some limited cleaning. File should be
a .txt in pipe-delimited ('|') format.
}
\seealso{
\code{\link[=clean_meds]{clean_meds()}}
}
