% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_mortality.R
\name{load_mortality}
\alias{load_mortality}
\title{load_mortality}
\usage{
load_mortality(
  mortality_filepath,
  coltypes_mortality = list(col_character(), col_datetime()),
  max_load = Inf
)
}
\arguments{
\item{mortality_filepath}{Path to the mortality .txt file}

\item{coltypes_mortality}{A list of cols() specifications. Cols specifications are things
like col_integer(), col_character(), and can be found within the 'readr' package documentation.}

\item{max_load}{A number, the maximum number of rows to load. The default is infinity.}
}
\value{
A data frame a list of MRNs and death dates
}
\description{
Load inpatient dates of death from a pipe-delimited .txt file.#'
}
