% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rass.R
\name{load_rass}
\alias{load_rass}
\title{load_rass}
\usage{
load_rass(
  rass_filepath,
  col_mapping = list(enc_id = "pat_enc_csn_id", mrn = "mrn", rass_id_col =
    "display_name", rass = "measure_value_raw", rass_time = "recorded_time"),
  rass_coltypes = NULL,
  max_load = Inf
)
}
\arguments{
\item{rass_filepath}{Path to the RASS data}

\item{col_mapping}{List that defines mapping of columns to key variables that will be output. You can dynamically
specify these mappings when calling the function.
The default is:
\itemize{
\item enc_id = "pat_enc_csn_id",
\item mrn = "mrn",
\item rass_id_col = "display_name",
\item rass = "measure_value_raw",
\item rass_time = "recorded_time"
}
The names on the left hand side of the list argument must be kept consistent.
The names on the right-hand side are the names that are actually present in the .txt file
which will be mapped to these variables.}

\item{rass_coltypes}{A list of cols() specifications specifying how columns should be read.
Cols specifications are things like col_integer(), col_character(), and can be found
within the \code{\link[readr]{cols}} documentation from the \code{readr} package.
By default it is recommended to just send in 'ccccccc' where the length of the string of
characters is the number of columns in the data.
If this isn't working well you can send in col_guess() for each one.}

\item{max_load}{The maximum number of rows to load. The default is \code{Inf}}
}
\value{
A data frame with:
\itemize{
\item \code{mrn}: Medical record number
\item \code{enc_id}: Encounter ID, renamed from PAT_ENC_CSN_ID
\item \code{rass_time}: Datetime when the RASS was recorded
\item \code{rass}: Integer value of RASS ranging from -5 to +4 representing:
\itemize{
\item -5 Unarousable
\item -4 Deep sedation
\item -3 Moderate sedation
\item -2 Light sedation
\item -1 Drowsy
\item 0 Alert and calm
\item 1 Restless
\item 2 Agitated
\item 3 Very agitated
\item 4 Combative
}
}
}
\description{
Load Richmond Agitation Sedation Scale data from Epic flowsheets.
The datafile contains other types of flowsheet data, such as delirium
(CAPD) scores. These are just filtered out. No processing is done
other than cleaning the dataset.
}
