% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_vent.R
\name{load_vent}
\alias{load_vent}
\title{load_vent}
\usage{
load_vent(
  vent_filepath,
  vent_coltypes = list(col_character(), col_character(), col_character(),
    col_character(), col_datetime()),
  vent_to_load = NA,
  max_load = Inf
)
}
\arguments{
\item{vent_filepath}{Path to the ventilator data}

\item{vent_coltypes}{A list of cols() specifications.
Cols specifications are things like col_integer(), col_character(), and can be found
within the \code{\link[readr]{cols}} documentation from the \code{readr} package.
If this isn't working well you can send in col_guess().}

\item{vent_to_load}{Names of vent settings/measurements to load. Defaults to all.}

\item{max_load}{Maximum number of rows}
}
\value{
A data frame with vent/respiratory support flowsheet rows.
}
\description{
Load in ventilator settings, measurements, and other related respiratory data
from a pipe-delimited .txt file. This includes things like BiPAP, CPAP, high flow,
nitric oxide, as well as some related measurements like SpO2. The only cleaning
that is done is ensuring settings are in the correct format.
}
