% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chonyepicdata-package.R
\docType{package}
\name{chonyepicdata-package}
\alias{chonyepicdata}
\alias{chonyepicdata-package}
\title{chonyepicdata: Loads, cleans, and processes CHONY EHR inpatient data from the Epic era (after 2/1/2023)}
\description{
This section TBD at a later date.
}
\keyword{internal}
