% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_meds.R
\name{clean_meds}
\alias{clean_meds}
\title{clean_meds}
\usage{
clean_meds(
  df_meds,
  medlist = c("midazolam", "lorazepam", "diazepam", "clonazepam", "alprazolam",
    "morphine", "fentanyl", "hydromorphone", "oxycodone", "clonidine", "ketamine",
    "pentobarbital", "quetiapine", "haloperidol", "risperidone", "olanzapine",
    "aripiprazole", "diphenhydramine", "hydroyxyzine", "rocuronium", "vecuronium",
    "cisatracurium", "dexmedetomidine"),
  mar_med_given = c("anesthesia volume adjustment",
    "bolus from bag (dual sign required)", "bolus from bag",
    "continue to inpatient floor", "continued from or", "continued from pre",
    "given by other", "given during downtime", "given", "handoff (dual sign required)",
    "handoff", "new bag", "new bag/syringe/cartridge", "override pull", "rate change",
    "rate verify", "rate/dose change", "rate/dose changed", "rate/dose verify",
    "rate/dose verify", "bolus", "restarted (dual sign required)", "restarted",
    "started during downtime", 
     "started", "verification"),
  mar_med_stopped = c("held", "held by provider", "mar hold",
    "stopped (dual sign required)", "stopped", "stop infusion"),
  row_limit = Inf,
  time_limits = NA,
  patient_weights = NULL
)
}
\arguments{
\item{df_meds}{Medication datafile. Must have columns with medication names,
doses, concentrations, and rates.}

\item{medlist}{An array of characters that will be matched to determine
medication names to keep. Partial matches can be used according to \link{stringr} rules.
Pass \code{NA} if you you don't want to filter at all.}

\item{mar_med_given}{An array of characters that specify whether a medication
was actually given to the patient. Default values are provided. Rows that
do not match this will be excluded.}

\item{mar_med_stopped}{An array of characters that specify if a medication was
definitively stopped. Default values are provided.}

\item{row_limit}{Maximum number of rows to allow. Useful for debugging very
large files.}

\item{time_limits}{Data table of encounter IDs and time intervals during
which medication doses should be included. The first column should be named
'enc_id' and the second column should be formatted as time intervals using \code{\link[lubridate:interval]{lubridate::interval()}}}

\item{patient_weights}{Data table of encounter IDs matched to a weight variable.
If provided, will adjust all medications to weight-based dosing. If not provided, will
separate output into a weight-based and non weight-based component for each medication.
This is because some medications only have a weight-based dose recorded, and not an
absolute dose. Other meds have an absolute dose. You can't combine these unless
you also have the weight.}
}
\value{
A data frame with raw medication data. Medication name is saved in character format
in the column 'med_name'. These should be cleaned and processed. If time limits were sent in via
\code{time_limits}, then this will be sorted by individual PICU stay. Returns weight-based doses if
weights were included in the function call. If not, it will include weight-based versus raw doses
(in mg, grams, units, etc) based upon the order. These can be converted to weight-based post-hoc.
The total duration of infusions in hours is returned as cumul_time variables.
}
\description{
Clean and process raw medication data loaded from Epic EHR. The majority of
this work is string matching and removal to extract meaningful medication
names, doses, and concentrations. Doses are not weight-based unless recorded
as such.
}
