% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_meds.R
\name{clean_meds}
\alias{clean_meds}
\title{clean_meds}
\usage{
clean_meds(
  df_meds,
  medlist = NULL,
  mar_med_given = c("anesthesia volume adjustment",
    "bolus from bag (dual sign required)", "bolus from bag",
    "continue to inpatient floor", "continued from or", "continued from pre",
    "given by other", "given during downtime", "given", "handoff (dual sign required)",
    "handoff", "new bag", "new bag/syringe/cartridge", "override pull", "rate change",
    "rate verify", "rate/dose change", "rate/dose changed", "rate/dose verify",
    "rate/dose verify", "bolus", "restarted (dual sign required)", "restarted",
    "started during downtime", 
     "started", "verification"),
  mar_med_stopped = c("held", "held by provider", "mar hold",
    "stopped (dual sign required)", "stopped", "stop infusion"),
  row_limit = Inf
)
}
\arguments{
\item{df_meds}{Medication datafile. Must have columns with medication names,
doses, concentrations, and rates.}

\item{medlist}{An array of characters that will be matched to determine
medication names to keep. Partial matches can be used according to \code{\link[stringr]} rules.
If argument is provided, no filtering will be performed at this level.}

\item{mar_med_given}{An array of characters that specify whether a medication
was actually given to the patient. Default values are provided. Rows that
do not match this will be excluded.}

\item{mar_med_stopped}{An array of characters that specify if a medication was
definitively stopped. Default values are provided.}

\item{row_limit}{Maximum number of rows to allow. Useful for debugging very
large files.}
}
\value{
A data frame with raw medication data. Medication name is saved in character format
in the column 'med_name'. These should be cleaned and processed.
}
\description{
Clean and process raw medication data loaded from Epic EHR. The majority of
this work is string matching and removal to extract meaningful medication
names, doses, and concentrations. Doses are not weight-based unless recorded
as such.
}
\seealso{
\code{\link[=clean_meds]{clean_meds()}}
}
