% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_resp_support_old.R
\name{clean_resp_support_old}
\alias{clean_resp_support_old}
\title{clean_resp_support_old}
\usage{
clean_resp_support_old(df_resp)
}
\arguments{
\item{df_resp}{A long-form data frame of respiratory settings and measurements}
}
\value{
A data frame in wide format with concurrently-recorded respiratory data
}
\description{
Take respiratory support file loaded by \link{load_resp_support}. Clean it a little,
keep standard valuable rows, and put it into wide format so that each timestamp has
all concurrently-recorded values.
}
\details{
NOTE: This is a deprecated file that is being phased out as of 7/18/2024. Data files that
occurred after this date will use clean_resp_support().
}
