% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_vitals.R
\name{clean_vitals}
\alias{clean_vitals}
\title{clean_vitals}
\usage{
clean_vitals(df_vitals)
}
\arguments{
\item{df_vitals}{A long-form data frame of vitals.}
}
\value{
A data frame in wide format with concurrently-recorded vital signs
}
\description{
Take vitals file loaded by \link{load_vitals} and standardize it so that
it only includes heart rate, BP (arterial and non-invasive), respirations,
SpO2, cvp. Data will be output in wide-format with vitals recorded at the same time
taking up the same row.
}
