% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coma_intervals.R
\name{get_coma_intervals}
\alias{get_coma_intervals}
\title{get_coma_intervals}
\usage{
get_coma_intervals(df_rass_intervals)
}
\arguments{
\item{df_rass_intervals}{A data frame returned by \link{get_rass_intervals}.}

\item{id}{A character vector of IDs for each unique patient or encounter. Generally taken from
the variable PAT_ENC_CSN_ID, but can also create custom IDs. An example would be to use
PAT_ENC_CSN_ID but also append an identifier (01, 02, 03, etc.) for each PICU hospitalization
or episode of invasive mechanical ventilation.}
}
\value{
A data frame with:
\itemize {
\item \code{id}: The ID of the patient.
\item \code{coma_time_start}: A POSIXct object when a coma episode started.
\item \code{coma_time_stop}: A POSIXct object when a coma episode ended.
}
}
\description{
Takes data fram from \link{get_rass_intervals} and creates intervals during which
each patient encounter was comatose (RASS -4 or -5).
}
