% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_imv_startstop.R
\name{get_imv_startstop}
\alias{get_imv_startstop}
\title{get_imv_startstop}
\usage{
get_imv_startstop(
  df_vent_wide,
  min_inter_ep_duration = 2,
  min_ep_duration = 24
)
}
\arguments{
\item{df_vent_wide}{A wide-format data frame of vent/respiratory settings and measurements. This should
be obtained using \link{clean_vent}}

\item{min_inter_ep_duration}{The minimum duration of time (in hours) between episodes of respiratory support
for them to be counted as separate. This is designed to make sure that erroneous entries that would flag a patient
as extubated will be removed. Neighboring valid episodes will be joined together.}

\item{min_ep_duration}{Minimum duration of time (in hours) to count as a period of continuous ventilation. Default
is 24 hours. Shorter episodes will be removed.}
}
\value{
A data frame in long format with start and stop times of IMV for each patient/encounter
}
\description{
Find times where invasive mechanical ventilation (IMV) starts and stops, to determine
the exact interval and duration of IMV.
}
