% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_capd.R
\name{load_capd}
\alias{load_capd}
\title{load_capd}
\usage{
load_capd(
  capd_filepath,
  capd_colpaths = list(col_character(), col_character(), col_skip(), col_character(),
    col_skip(), col_number(), col_skip(), col_skip(), col_character()),
  max_load = Inf
)
}
\arguments{
\item{capd_filepath}{Path to the CAP-D data}

\item{capd_colpaths}{A list of cols() specifications.
Cols specifications are things like col_integer(), col_character(), and can be found
within the \code{\link[readr]{cols}} documentation from the \code{readr} package.
If this isn't working well you can send in col_guess().}

\item{max_load}{The maximum number of rows to load. The default is \code{Inf}}
}
\value{
#' @return A data frame with:
\itemize{
\item \code{mrn}: Medical record number
\item \code{enc_id}: Encounter ID, renamed from PAT_ENC_CSN_ID
\item \code{capd_time}: Datetime when the CAP-D was recorded
\item \code{capd}: Integer value of the summed components of all CAP-D. A score
of 9 or more means the patient screened positive for delirium.
Any time where a component of the CAP-D was recorded as NA will not be returned.
Scores need to be complete.
}
}
\description{
Load Cornell Assessment of Pediatric Delirium (CAP-D) data from Epic flowsheets.
The datafile contains other types of flowsheet data, which are filtered out.
Columns must have the names defined in the dataset from November 2023 (for now).
}
