% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_encounters.R
\name{load_encounters}
\alias{load_encounters}
\title{Load hospital encounters}
\usage{
load_encounters(
  encounter_filepath,
  coltypes_enc = list(col_character(), col_character(), col_skip(), col_datetime(),
    col_character(), col_double(), col_skip(), col_skip(), col_datetime(),
    col_datetime()),
  max_load = Inf
)
}
\arguments{
\item{encounter_filepath}{A complete file path to the encounter data file.}

\item{coltypes_enc}{A list of cols() specifications.
Cols specifications are things like col_integer(), col_character(), and can be found within the 'readr' package documentation.}

\item{max_load}{A number, the maximum number of rows to load. The default is infinity.}
}
\value{
A data frame with reformatted and cleaned encounter-level information, including MRN, enc_id (encounter ID), dob (date of birth), and hospital admission and discharge dates and datetimes
}
\description{
Load inpatient hospital encounters from a .txt file from Epic-era data.
The text file should be pipe-delimited (delimter is '|' without the quotes).
This just contains basic demographic information lke dates os hospitalization,
age, and linking information to be used with other relational tables.
}
