% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_generic_flowsheet_rows.R
\name{load_generic_flowsheet_rows}
\alias{load_generic_flowsheet_rows}
\title{load_generic_flowsheet_rows}
\usage{
load_generic_flowsheet_rows(
  flowsheet_filepath,
  key_name = "PAT_ENC_CSN_ID",
  time_col = NULL,
  var_col = NULL,
  measure_col = NULL,
  varnames = NULL,
  rename_vars = NULL,
  max_load = Inf
)
}
\arguments{
\item{flowsheet_filepath}{Path to the flowsheet}

\item{key_name}{Name of column(s) containing unique linking identifiers}

\item{time_col}{Column with timestamps, typically in MDY_HMS format}

\item{var_col}{Column containing the names of variables. Should only have one.}

\item{measure_col}{Column with measurements corresponding to variables.}

\item{varnames}{The actual names of variables contained in \code{var_col}. Can
have multiple.}

\item{rename_vars}{What variable names from \code{varnames} should be renamed to. Must
match 1:1 with that variable.}

\item{max_load}{Maximum number of rows to load.}
}
\value{
A data frame with the identity key and timestamped data points from the file.
}
\description{
Send in a flowsheet file of an undefined type, and pull out the relevant rows
in a wide, timestamped format. This is because some flowsheets may have data that
requires pulling into a custom data frame witih one or several variables.
}
