% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_icd_dx.R
\name{load_icd_dx}
\alias{load_icd_dx}
\title{load_icd_dx}
\usage{
load_icd_dx(
  icd_dx_filepath,
  coltypes_icd = list(col_character(), col_character(), col_character(), col_character(),
    col_datetime(), col_character(), col_factor()),
  max_load = Inf
)
}
\arguments{
\item{icd_dx_filepath}{Path to the ICD diagnosis code file.}

\item{coltypes_icd}{A list of cols() specifications.
Cols specifications are things like col_integer(), col_character(), and can be found within the 'readr' package documentation.}

\item{max_load}{A number, the maximum number of rows to load. The default is infinity.}
}
\value{
A data frame with:
\itemize{
\item \code{mrn}: Medical record number
\item \code{enc_id}: Encounter ID, renamed from PAT_ENC_CSN_ID
\item \code{icd10_code}: Numeric ICD10 code with the format XXXXXXX (the period character is removed)
\item \code{dx_name}: Description of the ICD10 diagnosis
\item \code{dx_date}: Date the diagnosis was placed
\item \code{primary_dx_yn}: Is this a primary diagnosis (Y/N)? (Usually the one listed first)
\item \code{dx_type}: Where did the diagnosis come from? (Billing records, problem list, etc)
}
}
\description{
Use this function to load a set of ICD10 diagnosis codes. ICD codes will be
loaded from many different sources, and across time. Outside of this function they can be joined to
particular patients or encounters.
}
