% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_labs.R
\name{load_labs}
\alias{load_labs}
\title{load_labs}
\usage{
load_labs(
  labs_filepath,
  labs_coltypes = list(col_skip(), col_skip(), col_skip(), col_character(), col_skip(),
    col_datetime(), col_skip(), col_character(), col_character(), col_character(),
    col_character(), col_skip()),
  max_load = Inf
)
}
\arguments{
\item{labs_filepath}{Path to the lab data}

\item{labs_coltypes}{A list of cols() specifications.
Cols specifications are things like col_integer(), col_character(), and can be found
within the \code{\link[readr]{cols}} documentation from the \code{readr} package.
If this isn't working well you can send in col_guess().}

\item{max_load}{The maximum number of rows to load. The default is \code{Inf}}
}
\value{
A data frame with raw laboratory data and timestamps.
}
\description{
Load laboratory data from the Epic EHR, and perform some limited cleaning. File should be
a .txt in pipe-delimited ('|') format.
}
