% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rass.R
\name{load_rass}
\alias{load_rass}
\title{load_rass}
\usage{
load_rass(
  rass_filepath,
  rass_colpaths = list(col_character(), col_character(), col_skip(), col_character(),
    col_skip(), col_number(), col_skip(), col_skip(), col_character()),
  max_load = Inf
)
}
\arguments{
\item{rass_filepath}{Path to the RASS data}

\item{rass_colpaths}{A list of cols() specifications.
Cols specifications are things like col_integer(), col_character(), and can be found within the 'readr' package documentation.}

\item{max_load}{A number, the maximum number of rows to load. The default is infinity.}
}
\value{
A data frame with:
\itemize{
  \item \code{mrn}: Medical record number
  \item \code{enc_id}: Encounter ID, renamed from PAT_ENC_CSN_ID
  \item \code{rass_time}: Datetime when the RASS was recorded
  \item \code{rass}: Integer value of RASS [-5, 4] representing
   &nbsp;&nbsp;-5 Unarousable
   &nbsp;&nbsp;-4 Deep sedation
   &nbsp;&nbsp;-3 Moderate sedation
   &nbsp;&nbsp;-2 Light sedation
   &nbsp;&nbsp;-1 Drowsy
   &nbsp;&nbsp;0 Alert and calm
   &nbsp;&nbsp;1 Restless
   &nbsp;&nbsp;2 Agitated
   &nbsp;&nbsp;3 Very agitated
   &nbsp;&nbsp;4 Combative
}
}
\description{
Load Richmond Agitation Sedation Scale data from Epic flowsheets.
The datafile contains other types of flowsheet data, such as delirium
(CAPD) scores. These are just filtered out. No processing is done
other than cleaning the dataset.
}
