% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rass.R
\name{load_rass}
\alias{load_rass}
\title{load_rass}
\usage{
load_rass(
  rass_filepath,
  rass_coltypes = list(col_character(), col_character(), col_skip(), col_character(),
    col_skip(), col_number(), col_skip(), col_skip(), col_character()),
  max_load = Inf
)
}
\arguments{
\item{rass_filepath}{Path to the RASS data}

\item{rass_coltypes}{A list of cols() specifications.
Cols specifications are things like col_integer(), col_character(), and can be found
within the \code{\link[readr]{cols}} documentation from the \code{readr} package.
If this isn't working well you can send in col_guess().}

\item{max_load}{The maximum number of rows to load. The default is \code{Inf}}
}
\value{
A data frame with:
\itemize{
\item \code{mrn}: Medical record number
\item \code{enc_id}: Encounter ID, renamed from PAT_ENC_CSN_ID
\item \code{rass_time}: Datetime when the RASS was recorded
\item \code{rass}: Integer value of RASS ranging from -5 to +4 representing:
\itemize{
\item -5 Unarousable
\item -4 Deep sedation
\item -3 Moderate sedation
\item -2 Light sedation
\item -1 Drowsy
\item 0 Alert and calm
\item 1 Restless
\item 2 Agitated
\item 3 Very agitated
\item 4 Combative
}
}
}
\description{
Load Richmond Agitation Sedation Scale data from Epic flowsheets.
The datafile contains other types of flowsheet data, such as delirium
(CAPD) scores. These are just filtered out. No processing is done
other than cleaning the dataset.
}
