% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_vitals.R
\name{load_vitals}
\alias{load_vitals}
\title{load_vitals}
\usage{
load_vitals(
  vitals_filepath,
  vitals_coltypes = list(col_character(), col_character(), col_character(),
    col_character(), col_character(), col_character(), col_character(), col_character(),
    col_datetime()),
  vitals_to_load = NA,
  max_load = Inf
)
}
\arguments{
\item{vitals_filepath}{Path to the medication data}

\item{vitals_coltypes}{A list of cols() specifications.
Cols specifications are things like col_integer(), col_character(), and can be found
within the \code{\link[readr]{cols}} documentation from the \code{readr} package.
If this isn't working well you can send in col_guess().}

\item{vitals_to_load}{Names of vital signs to load, in array format}

\item{max_load}{Maximum number of rows}
}
\value{
A data frame with vital signs (and other related flowsheet data) loaded.
}
\description{
load_vitals
}
